Fixed for v41.53 [IWBUMS]
Updated the PluginTemplate.lua file to have the latest version of the method as an example of what is going on behind the scenes. This way, other modders can decide better on how to modify the behavior without digging through the source files.

Also cleaned up some functionality to make it more verbose, as well as adding more documentation to code that isn't as easy to understand at a glance (especially the PluginTemplate)

I had to completely change how purses are loaded with money because the latest update changes the loading behavior away from the OnFillContainer event and it doesn't seem to be directly accessible right now. As a workaround I have shifted to the OnZombieDead event. With it I check if the zombie is female and using a new config option, there's a chance to give a purse with money in it.

The snacks and drinks to be loaded can now be configured as well, including ones from other mods. They can be defined in the config directly, or mods can modify the config options when they load.

When ZombRand is being invoked, the max value is increased by 1 in order to make it inclusive instead of exclusive. (Range changed from [min,max) to [min,max])

The spawn chance for wallets is now an option in the config. It overrides the new spawn chances into the distribution table, so it's not directly based off percentages and I don't actually know exactly how it works. I merely doubled the spawn chance as a base in order to make money show up just a bit more so it's actually useful and doesn't run dry instantly.